requirements.txt
amqp==5.0.9
autopep8==1.6.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.11
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
dnspython==1.16.0
eventlet==0.31.1
Flask==2.0.2
Flask-HTTPAuth==4.5.0
flask-paginate==2022.1.8
flower==1.0.0
greenlet==1.1.2
humanize==3.14.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpickle==1.2
kombu==5.2.3
MarkupSafe==2.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.27
pycodestyle==2.8.0
pymongo==3.12.0
pytz==2021.3
redis==3.5.3
redisbeat==1.2.4
requests==2.27.1
six==1.16.0
toml==0.10.2
tornado==6.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
python-slugify
flask-login